add basic auth

+++ tests/web/admin/test_admin_ingredients_controller.py
@@ -3,56 +3,64 @@ from db import session
 from schemas import IngredientClientSchema
 
 
-def test_status(test_client):
+def test_status_without_auth(test_client):
     response = test_client.get('/api/admin/v1/ingredients/')
+    actual = response.status_code
+    expected = 401
+
+    assert actual == expected
+
+
+def test_status(test_client, admin_key):
+    response = test_client.get('/api/admin/v1/ingredients/', headers={'Authorization': f'Basic {admin_key}'})
     actual = response.status_code
     expected = 200
 
     assert actual == expected
 
 
-def test_count(test_client):
+def test_count(test_client, admin_key):
     ingredient = Ingredient(name='Name', calories=200)
     session.add(ingredient)
     session.commit()
 
-    response = test_client.get('/api/admin/v1/ingredients/')
+    response = test_client.get('/api/admin/v1/ingredients/', headers={'Authorization': f'Basic {admin_key}'})
     actual = len(response.json['result'])
     expected = 1
 
     assert actual == expected
 
 
-def test_get(test_client):
+def test_get(test_client, admin_key):
     ingredient = Ingredient(name='Name', calories=200)
     session.add(ingredient)
     session.commit()
 
-    response = test_client.get('/api/admin/v1/ingredients/')
+    response = test_client.get('/api/admin/v1/ingredients/', headers={'Authorization': f'Basic {admin_key}'})
     actual = response.json['result'][0]
     expected = IngredientClientSchema().dump(ingredient)
 
     assert actual == expected
 
 
-def test_show(test_client):
+def test_show(test_client, admin_key):
     ingredient = Ingredient(name='Name', calories=200)
     session.add(ingredient)
     session.commit()
 
-    response = test_client.get(f"/api/admin/v1/ingredients/{ingredient.id}/")
+    response = test_client.get(f"/api/admin/v1/ingredients/{ingredient.id}/", headers={'Authorization': f'Basic {admin_key}'})
     actual = response.json['result']
     expected = IngredientClientSchema().dump(ingredient)
 
     assert actual == expected
 
 
-def test_create(test_client):
+def test_create(test_client, admin_key):
     data = ({
         "name": "Name",
         "calories": 200,
     })
-    response = test_client.post('/api/admin/v1/ingredients/', json=data)
+    response = test_client.post('/api/admin/v1/ingredients/', json=data, headers={'Authorization': f'Basic {admin_key}'})
 
     actual = response.json['result']
     expected = IngredientClientSchema().dump(session.query(Ingredient).one())
@@ -60,7 +68,7 @@ def test_create(test_client):
     assert actual == expected
 
 
-def test_update(test_client):
+def test_update(test_client, admin_key):
     ingredient = Ingredient(name='Name', calories=200)
     session.add(ingredient)
     session.commit()
@@ -69,7 +77,7 @@ def test_update(test_client):
         "name": "Name111",
         "calories": 2001,
     })
-    response = test_client.put(f"/api/admin/v1/ingredients/{ingredient.id}/", json=data)
+    response = test_client.put(f"/api/admin/v1/ingredients/{ingredient.id}/", json=data, headers={'Authorization': f'Basic {admin_key}'})
 
     actual = response.json['result']
     expected = IngredientClientSchema().dump(ingredient)
@@ -77,12 +85,12 @@ def test_update(test_client):
     assert actual == expected
 
 
-def test_delete(test_client):
+def test_delete(test_client, admin_key):
     ingredient = Ingredient(name='Name', calories=200)
     session.add(ingredient)
     session.commit()
 
-    test_client.delete(f"/api/admin/v1/ingredients/{ingredient.id}/")
+    test_client.delete(f"/api/admin/v1/ingredients/{ingredient.id}/", headers={'Authorization': f'Basic {admin_key}'})
     actual = session.query(Ingredient).count()
     expected = 0
