+++ tests/web/admin/test_admin_receipts_controller.py
@@ -4,15 +4,23 @@ from schemas import ReceiptClientSchema
 from tests.test_receipts import prepare_create_receipt
 
 
-def test_status(test_client):
+def test_status_without_auth(test_client):
     response = test_client.get('/api/admin/v1/receipts/')
+    actual = response.status_code
+    expected = 401
+
+    assert actual == expected
+
+
+def test_status(test_client, admin_key):
+    response = test_client.get('/api/admin/v1/receipts/', headers={'Authorization': f'Basic {admin_key}'})
     actual = response.status_code
     expected = 200
 
     assert actual == expected
 
 
-def test_count(test_client):
+def test_count(test_client, admin_key):
     first_ingredient, second_ingredient, first_step, second_step = prepare_create_receipt()
 
     receipt = Receipt(
@@ -26,14 +34,14 @@ def test_count(test_client):
     session.add(receipt)
     session.commit()
 
-    response = test_client.get('/api/admin/v1/receipts/')
+    response = test_client.get('/api/admin/v1/receipts/', headers={'Authorization': f'Basic {admin_key}'})
     actual = len(response.json['result'])
     expected = 1
 
     assert actual == expected
 
 
-def test_get(test_client):
+def test_get(test_client, admin_key):
     first_ingredient, second_ingredient, first_step, second_step = prepare_create_receipt()
 
     receipt = Receipt(
@@ -47,14 +55,14 @@ def test_get(test_client):
     session.add(receipt)
     session.commit()
 
-    response = test_client.get('/api/admin/v1/receipts/')
+    response = test_client.get('/api/admin/v1/receipts/', headers={'Authorization': f'Basic {admin_key}'})
     actual = response.json['result'][0]
     expected = ReceiptClientSchema().dump(receipt)
 
     assert actual == expected
 
 
-def test_show(test_client):
+def test_show(test_client, admin_key):
     first_ingredient, second_ingredient, first_step, second_step = prepare_create_receipt()
 
     receipt = Receipt(
@@ -68,14 +76,14 @@ def test_show(test_client):
     session.add(receipt)
     session.commit()
 
-    response = test_client.get(f"/api/admin/v1/receipts/{receipt.id}/")
+    response = test_client.get(f"/api/admin/v1/receipts/{receipt.id}/", headers={'Authorization': f'Basic {admin_key}'})
     actual = response.json['result']
     expected = ReceiptClientSchema().dump(receipt)
 
     assert actual == expected
 
 
-def test_create(test_client):
+def test_create(test_client, admin_key):
     first_ingredient, second_ingredient, first_step, second_step = prepare_create_receipt()
 
     data = ({
@@ -86,14 +94,14 @@ def test_create(test_client):
         "steps": [first_step, second_step],
     })
 
-    response = test_client.post('/api/admin/v1/receipts/', json=data)
+    response = test_client.post('/api/admin/v1/receipts/', json=data, headers={'Authorization': f'Basic {admin_key}'})
     actual = response.json['result']
     expected = ReceiptClientSchema().dump(session.query(Receipt).one())
 
     assert actual == expected
 
 
-def test_update(test_client):
+def test_update(test_client, admin_key):
     first_ingredient, second_ingredient, first_step, second_step = prepare_create_receipt()
 
     receipt = Receipt(
@@ -115,14 +123,14 @@ def test_update(test_client):
         "steps": [first_step],
     })
 
-    response = test_client.put(f"/api/admin/v1/receipts/{receipt.id}/", json=data)
+    response = test_client.put(f"/api/admin/v1/receipts/{receipt.id}/", json=data, headers={'Authorization': f'Basic {admin_key}'})
     actual = response.json['result']
     expected = ReceiptClientSchema().dump(receipt)
 
     assert actual == expected
 
 
-def test_delete(test_client):
+def test_delete(test_client, admin_key):
     first_ingredient, second_ingredient, first_step, second_step = prepare_create_receipt()
 
     receipt = Receipt(
@@ -136,7 +144,7 @@ def test_delete(test_client):
     session.add(receipt)
     session.commit()
 
-    test_client.delete(f"/api/admin/v1/receipts/{receipt.id}/")
+    test_client.delete(f"/api/admin/v1/receipts/{receipt.id}/", headers={'Authorization': f'Basic {admin_key}'})
     actual = session.query(Receipt).count()
     expected = 0
